Begin to modify channel structure to be less Java-oriented
This should allow for clearer and more reliable code, given that this
is Python and not Java. The idea is that strong typing doesn't exist in
references, so the Java technique of having inheritance for IO doesn't
really provide any guarantees. This system involves centralizing all
Input implementations into authoritative implementations, and making
Output implementations work based on supplied references.